추가: 우승자는 한 명 이상일 수 있다.
